REFACTOR: Rename tmp ts file
